auth/requirements.txt
flask==3.1.2
google-auth==2.41.1
google-auth-oauthlib==1.2.2
jwt==1.4.0
psycopg2==2.9.10